+++ sites/docs/src/lib/components/demos/element-rect.svelte
@@ -1,7 +1,9 @@
 <script lang="ts">
 	import { ElementRect } from "runed";
+	import DemoContainer from "$lib/components/demo-container.svelte";
+	import { Textarea } from "$lib/components/ui/textarea";
 
-	let el = $state() as HTMLElement;
+	let el = $state() as HTMLTextAreaElement;
 	const rect = new ElementRect(() => el);
 	const text = $derived(
 		Object.entries(rect.current)
@@ -10,11 +12,11 @@
 	);
 </script>
 
-<div class="rounded-md bg-card p-8">
-	<textarea
-		bind:this={el}
-		class="h-[330px] min-h-[300px] w-[300px] min-w-[200px] resize rounded-md bg-muted p-4"
+<DemoContainer>
+	<Textarea
+		bind:el
+		class="h-[330px] min-h-[300px] w-[300px] min-w-[200px] resize text-base"
 		value={text}
 		readonly
-	></textarea>
-</div>
+	/>
+</DemoContainer>