+++ sites/docs/src/lib/components/demos/use-debounce.svelte
@@ -1,5 +1,7 @@
 <script lang="ts">
 	import { useDebounce } from "runed";
+	import DemoContainer from "$lib/components/demo-container.svelte";
+	import { Button } from "$lib/components/ui/button";
 
 	let count = $state(0);
 	let logged = $state("");
@@ -21,11 +23,7 @@
 	}
 </script>
 
-<div class="rounded-md bg-card p-8">
-	<button
-		class="inline-flex items-center justify-center rounded-md border bg-brand/50
-		px-3 py-1.5 text-sm font-medium transition-all hover:bg-brand/25 focus:outline-none active:bg-brand/15"
-		onclick={ding}>DING DING DING</button
-	>
+<DemoContainer>
+	<Button variant="brand" size="sm" onclick={ding}>DING DING DING</Button>
 	<p>{logged || "Press the button!"}</p>
-</div>
+</DemoContainer>