+++ sites/docs/src/lib/components/demos/use-event-listener.svelte
@@ -1,10 +1,11 @@
 <script lang="ts">
 	import { useEventListener } from "runed";
+	import DemoContainer from "$lib/components/demo-container.svelte";
 
 	let count = $state(0);
 	useEventListener(document.body, "click", () => count++);
 </script>
 
-<div class="select-none rounded-md bg-card p-8">
+<DemoContainer class="select-none">
 	<p>You've clicked the document {count} {count === 1 ? "time" : "times"}</p>
-</div>
+</DemoContainer>