+++ packages/runed/CHANGELOG.md
@@ -1,5 +1,21 @@
 # runed
 
+## 0.10.0
+
+### Minor Changes
+
+- feat: useIntersectionObserver ([#73](https://github.com/svecosystem/runed/pull/73))
+
+### Patch Changes
+
+- fix: ensure "add" utility exports types as well ([#74](https://github.com/svecosystem/runed/pull/74))
+
+## 0.9.1
+
+### Patch Changes
+
+- rename match to matches in MediaQuery ([#69](https://github.com/svecosystem/runed/pull/69))
+
 ## 0.9.0
 
 ### Minor Changes